Refactor translation package layout
