+++ nemo_curator/stages/text/translation/__init__.py
@@ -1,40 +1,24 @@
-# Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from nemo_curator.stages.text.translation.faith_eval import FaithEvalFilter
-from nemo_curator.stages.text.translation.field_utils import (
+from nemo_curator.stages.text.translation.evaluation.faith import FaithEvalFilter
+from nemo_curator.stages.text.translation.evaluation.text_quality import (
+    TextQualityMetricStage,
+    compute_text_quality_metric,
+)
+from nemo_curator.stages.text.translation.pipeline import TranslationPipeline
+from nemo_curator.stages.text.translation.stages.reassembly import ReassemblyStage
+from nemo_curator.stages.text.translation.stages.segmentation import SegmentationStage
+from nemo_curator.stages.text.translation.stages.translate import TranslateStage
+from nemo_curator.stages.text.translation.utils.field_paths import (
     extract_nested_fields,
     is_wildcard_path,
     normalize_text_field,
     parse_structured_value,
     set_nested_fields,
 )
-from nemo_curator.stages.text.translation.output_utils import (
+from nemo_curator.stages.text.translation.utils.metadata import (
     build_translation_metadata,
     merge_faith_scores_into_metadata,
     reconstruct_messages_with_translation,
 )
-from nemo_curator.stages.text.translation.pipeline import (
-    TranslationPipeline,
-)
-from nemo_curator.stages.text.translation.round_trip_eval import (
-    TextQualityMetricStage,
-    compute_text_quality_metric,
-)
-from nemo_curator.stages.text.translation.reassembly import ReassemblyStage
-from nemo_curator.stages.text.translation.segmentation import SegmentationStage
-from nemo_curator.stages.text.translation.translate import TranslateStage
 
 __all__ = [
     "FaithEvalFilter",
