+++ nemo_curator/stages/text/translation/utils/__init__.py
@@ -1,0 +1,27 @@
+from nemo_curator.stages.text.translation.utils.async_utils import run_async_safe
+from nemo_curator.stages.text.translation.utils.field_paths import (
+    extract_nested_fields,
+    is_wildcard_path,
+    normalize_text_field,
+    parse_structured_value,
+    set_nested_fields,
+)
+from nemo_curator.stages.text.translation.utils.metadata import (
+    build_translation_metadata,
+    merge_faith_scores_into_metadata,
+    reconstruct_messages_with_translation,
+)
+from nemo_curator.stages.text.translation.utils.prompt_loader import load_prompt_template
+
+__all__ = [
+    "build_translation_metadata",
+    "extract_nested_fields",
+    "is_wildcard_path",
+    "load_prompt_template",
+    "merge_faith_scores_into_metadata",
+    "normalize_text_field",
+    "parse_structured_value",
+    "reconstruct_messages_with_translation",
+    "run_async_safe",
+    "set_nested_fields",
+]
