+++ nemo_curator/stages/text/utils/__init__.py
@@ -1,0 +1,3 @@
+from nemo_curator.stages.text.utils.text_utils import get_language_name
+
+__all__ = ["get_language_name"]
