+++ tests/stages/text/translation/test_reassembly.py
@@ -21,8 +21,11 @@ import json
 import pandas as pd
 import pytest
 
-from nemo_curator.stages.text.translation.reassembly import ReassemblyStage, _INTERNAL_COLUMNS
-from nemo_curator.stages.text.translation.segmentation import SegmentationStage
+from nemo_curator.stages.text.translation.stages.reassembly import (
+    _INTERNAL_COLUMNS,
+    ReassemblyStage,
+)
+from nemo_curator.stages.text.translation.stages.segmentation import SegmentationStage
 from nemo_curator.tasks import DocumentBatch
